Updated user_edit.php with delete confirmation

+++ user/user_edit.php
@@ -34,7 +34,7 @@
         <td><?php echo $row["username"]; ?></td>
         <td><?php echo $row["email"]; ?></td>
         <td><a href="edit.php?id=<?=$row["user_id"];?>">Edit</a></td>
-        <td><a href="delete.php?id=<?=$row["user_id"];?>">Delete</a></td>
+        <td><a href="delete.php?id=<?=$row["user_id"];?>" onclick="return confirm('Apakah benar anda mau delete?');">Delete</a></td>
     </tr>
     <?php endwhile; ?>  
 </table>
